refactor: move PaymentGateway out of contracts namespace
